Updated `Array::shuffle()` to use `SWAP` instead of copying through temporary variables, which makes shuffling simple types around 2x faster, and refcounted object types around 20x faster.

Also updated multiple methods that insert into/modify the array to move their validated `Variant` to avoid an extra copy, speed increase varies depending on type and how much else the methods are doing, usually around 5-10% faster.

Compared with gdscript below:

```gdscript
func _ready() -> void:
	shuffle_int()
	shuffle_obj()
	insert_obj()
	push_string()
	set_index_int()
	set_index_str()
	set_obj()
	fill_str()

func shuffle_int():
	var a := range(1000000)
	var start := Time.get_ticks_msec()
	a.shuffle()
	var end := Time.get_ticks_msec()
	print("shuffle_int: %dms" % [end - start])

func shuffle_obj():
	var a := []
	for i in 1000000:
		a.push_back(TestClass.new())
		
	var start := Time.get_ticks_msec()
	a.shuffle()
	var end := Time.get_ticks_msec()
	print("shuffle_obj: %dms" % [end - start])

func insert_obj():
	var src := range(25000).map(func(i): return TestClass.new())
	var a := []
	var start := Time.get_ticks_msec()
	for i in src.size():
		a.insert(i / 2, src[i])
	var end := Time.get_ticks_msec()
	print("insert_obj: %dms" % [end - start])

func push_string():
	var src := range(1000000).map(str)
	var a := []
	var start := Time.get_ticks_msec()
	for i in src.size():
		a.push_back(src[i])
	var end := Time.get_ticks_msec()
	print("push_string: %dms" % [end - start])

func set_index_int():
	var src := range(1000000)
	var dst := []
	dst.resize(src.size())
	var start := Time.get_ticks_msec()
	for i in src.size():
		dst[i] = src[i]
	var end := Time.get_ticks_msec()
	print("set_index_int: %dms" % [end - start])

func set_index_str():
	var src := range(1000000).map(str)
	var dst := []
	dst.resize(src.size())
	var start := Time.get_ticks_msec()
	for i in src.size():
		dst[i] = src[i]
	var end := Time.get_ticks_msec()
	print("set_index_str: %dms" % [end - start])
	
func set_obj():
	var src := range(1000000).map(func(i): return TestClass.new())
	var dst := []
	dst.resize(src.size())
	var start := Time.get_ticks_msec()
	for i in src.size():
		dst.set(i, src.get(i))
	var end := Time.get_ticks_msec()
	print("set_obj: %dms" % [end - start])

func fill_str():
	var a := []
	a.resize(5)
	var start := Time.get_ticks_msec()
	for i in 1000000:
		a.fill(str(i))
	var end := Time.get_ticks_msec()
	print("fill_str: %dms" % [end - start])
	
class TestClass:
	var i : int
	var f : float
	var s : String
```

Old:
```
shuffle_int: 24ms
shuffle_obj: 303ms
insert_obj: 158ms
push_string: 184ms
set_index_int: 57ms
set_index_str: 97ms
set_obj: 289ms
fill_str: 442ms
```

New:
```
shuffle_int: 14ms
shuffle_obj: 14ms
insert_obj: 143ms
push_string: 160ms
set_index_int: 51ms
set_index_str: 80ms
set_obj: 250ms
fill_str: 419ms
```
<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->